meson: Use isolated install for integration tests
Make a separate install into the build dir for integration tests and
introduce GLYCIN_DATA_DIR to force glycin to use only this dir to lookup
configs.

This approach does not require configuring a specific writable prefix for
developers to be able to just run integration tests. Using
GLYCIN_DATA_DIR also ensures that no other installation is accidentally
used in the test.

Co-authored-by: Hubert Figuière <hub@figuiere.net>